src/requirements.txt
fastapi>=0.104.0
mangum>=0.17.0

python-dotenv>=1.0.0